requirements.txt
absl-py==0.12.0
astunparse==1.6.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
cycler==0.10.0
Flask==1.1.2
flatbuffers==1.12
gast==0.4.0
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
idna==2.10
importlib-metadata==3.10.0
itsdangerous==2.0.1
Jinja2==3.0.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-vggface==0.6
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
mtcnn==0.1.0
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.2.52
opt-einsum==3.3.0
Pillow==8.2.0
pip==21.1.1
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.3
setuptools==52.0.0.post20210125
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.5
Werkzeug==1.0.1
wheel==0.36.2
wrapt==1.12.1
zipp==3.4.1
